Deleted duplicate file, fixed other issues
